+++ src/components/WalletWidget/useShowTopUp.ts
@@ -1,0 +1,7 @@
+import { useCallback } from 'react';
+export const useShowTopUp = () => {
+
+  return useCallback(() => {
+    window.open('/home/topup');
+  }, []);
+};